requirements.txt
numpy==1.22.2
pandas==1.4.1
scipy==1.6.0
statsmodels==0.13.2
scikit-learn==0.24.0
matplotlib==3.5.1
adtk==0.6.2
PyYAML==6.0
setproctitle==1.2.1
pyarrow==7.0.0
